// types/auction-types.ts
export type ProductAttribute = {
  id: string;
  name: string;
  value: string;
  type: "text" | "number" | "select" | "color" | "boolean";
  required: boolean;
  options?: string[]; // For select type attributes
};

export type ProductCategory = {
  id: string;
  name: string;
  handle?: string;
  parentId?: string;
  level: number;
  subcategories?: ProductCategory[];
};